mortgage_calculator/requirements.txt
Flask==2.3.3
pandas==2.1.1
numpy==1.25.2
plotly==5.16.1
python-dateutil==2.8.2
Jinja2==3.1.2
Werkzeug==2.3.7